--- lib/pages/emergency_number_data.dart
@@ -1,40 +1,0 @@
-
-class EmergencyNumbersData {
-
-  static final data = [
-    {
-      'name': 'EMS-1',
-      'contact': '030000000',
-    },
-    {
-      'name': 'EMS-2',
-      'contact': '030000000',
-    },
-    {
-      'name': 'EMS-3',
-      'contact': '030000000',
-    },
-    {
-      'name': 'EMS-4',
-      'contact': '030000000',
-    },
-    {
-      'name': 'LUMS',
-      'contact': '030000000',
-      'gender': 'M',
-    },
-    {
-      'name': 'HAWC',
-      'contact': '030000000',
-    }
-  ];
-
-  // static final hello = [
-  //   {'text': 'Hello 1.', 'item': '1'},
-  //   {'text': 'Hello 2', 'item': '2'},
-  //   {'text': 'Hello 3', 'item': '3'},
-  //   {'text': 'Uhh nice', 'item': '4'},
-  //   {'text': 'Hello', 'item': '5'},
-  //   {'text': 'What', 'item': '6'}
-  // ];
-}
