+++ lib/pages/emergency_numbers.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:ems_direct/pages/emergency_number_data.dart';
 import 'package:ems_direct/pages/emergency_numbers_card.dart';
